Update plugin.py
When different commands were triggered one after the other, serveral ones were refused because the Nest commands were still in process. This version implements a buffer mechanism so that command are buffered and the next one is executed when the previous one is finished. Could be useful when triggering the devices in scripts (eg set off eco mode and start heating).